transactions amount works as integers instead of float

+++ database/migrations/2019_04_09_133315_create_transaction_table.php
@@ -19,7 +19,7 @@ class CreateTransactionTable extends Migration
             $table->unsignedBigInteger('user_to');
             $table->string('subject');
             $table->string('comment')->nullable();
-            $table->float('quantity');
+            $table->double('quantity');
             $table->timestamps();
 
             $table->foreign('user_from')->references('id')->on('users')->onDelete('cascade');
